Add files via upload
The portfolio project is centered around analyzing COVID-19 data, encompassing a wide range of statistical evaluations to understand the pandemic's impact globally and by specific locations, such as countries and continents. It leverages SQL queries to extract and compute critical metrics, including total cases, deaths, infection rates relative to population, and vaccination progress. Key aspects of the project include:

- **Comparative Analysis**: Evaluating total cases against total deaths to derive mortality rates and understand the severity of the pandemic across different regions.
- **Demographic Impact**: Calculating the percentage of the population affected by COVID-19, providing insights into how widespread the virus has been in various locations.
- **Geographical Focus**: Special attention is given to the United States, with queries tailored to assess the death percentage and infection rates within the country.
- **Infection and Mortality Rates**: Identifying countries and continents with the highest infection rates and death counts, highlighting regions most impacted by the pandemic.
- **Vaccination Progress**: Analyzing vaccination data through advanced SQL techniques, including joining tables, calculating rolling sums, and creating views to track vaccination rates over time.

This project employs sophisticated SQL operations, such as data partitioning, aggregation, and the creation of temporary tables and views, to offer a detailed and nuanced perspective on the COVID-19 pandemic's dynamics. It aims to provide actionable insights through the careful examination of case, death, and vaccination data, emphasizing the pandemic's health implications and the importance of vaccination efforts worldwide.