Add files via upload
The aim of this analysis is to determine the average income of gender, marital status that purchased bike or not, the commute distance of people that purchased bike or not, and the age bracket of people who purchased or not. The datasets contains information such as; ID, Marital Status, Gender , Income, Children, Education, Occupation, Home Owner, Cars, Commute Distance Region, Age, Purchased Bike. Excel was used for this analysis. Lets get right into it !

Step 1 : The dataset was imported into excel sheet. The data was gotten from alex_the_analyst bootcamp on youtube.
Step 2: The data was cleaned to remove duplicates and errors. 26 duplicate values found and removed.
Step 3: To give a more defined meaning to the ‘M’, ‘F’, and ‘S’, ctrl + h (find and replace) tool was used to replace them to ‘Married’, ‘Male’, ‘Female’, and single.

Step 4 : Grouping the ages in column L into age bracket in order to give a defined meaning. using the IF statement, I was able to group it into adolescent, middle age, and old.

=IF(L2>54,”Old”,IF(L2>=31,”Middle Age”, IF(L2 < 31,”Adolescent”,”Invalid”)))
Step 4 : The dataset was insert into pivot table. Pivot table is a tool use for summarizing and analyzing data.
Step 5: Building the bike sales dashboard. Dashboards are visual representations of data.
Findings: The dashboard shows that the male gender and married people had the highest income and they also had the highest purchased bike. For the commute distance, 0–1 miles commuters purchased more bikes. Ages between 31–53(middle age) also had the highest number of purchase.